# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/videos": {
      "post": {
        "operationId": "VideosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save a video link. 201 when created, 200 when this id was already saved.",
        "tags": [
          "videos"
        ]
      },
      "get": {
        "operationId": "VideosController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The library, newest first.",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/v1/videos/{id}": {
      "patch": {
        "operationId": "VideosController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename a video.",
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "operationId": "VideosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a video from the library.",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/v1/videos/{id}/playback": {
      "put": {
        "operationId": "VideosController_playback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remember where the viewer stopped, for every device.",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "VideoBridge API",
    "description": "REST API for the VideoBridge phone and TV apps. Every non-2xx response uses the error envelope.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVideoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Client-generated UUID; makes the request safe to retry."
          },
          "sourceUrl": {
            "type": "string",
            "example": "https://cdn.example.com/a/My.Holiday.2024.mp4"
          },
          "title": {
            "type": "string",
            "description": "Defaults to a title derived from the link."
          },
          "sizeBytes": {
            "type": "number",
            "description": "File size in bytes, if the saving device could read it."
          },
          "format": {
            "type": "string",
            "example": "MP4",
            "description": "Short container/stream name for display."
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "A picture the link advertises (og:image). http(s) only."
          }
        },
        "required": [
          "sourceUrl"
        ]
      },
      "RenameVideoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "PlaybackDto": {
        "type": "object",
        "properties": {
          "positionMs": {
            "type": "number",
            "description": "Where the viewer stopped, in milliseconds."
          },
          "durationMs": {
            "type": "number"
          }
        },
        "required": [
          "positionMs"
        ]
      }
    }
  }
}